fixed magic recharge... really wouldve expected a compile error from the gem 'int amount = 20.f' and a bound function that expects a float input but whatever

+++ Ryddelmyst/Source/Ryddelmyst/Private/RyddelmystCharacter.cpp
@@ -55,7 +55,7 @@ void ARyddelmystCharacter::BeginPlay()
 
 	// linear magic recharge; recharges in 20 point blocks every 5 seconds
 	FTimerDelegate TimerDelegate;
-	int MagicRechargeAmount = 20.f;
+	float MagicRechargeAmount = 20.f;
 	TimerDelegate.BindUFunction(this, FName("UpdateMagic"), MagicRechargeAmount);
 	GetWorldTimerManager().SetTimer(MagicTimerHandle, TimerDelegate, 5.0f, true, 0.f);
